requirements.txt
Django==5.0.2
channels==4.0.0
channels-redis==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
redis==5.0.1 